Rename the controls.
