+++ DiffPlex.Wpf/Controls/Helper.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace DiffPlex.Wpf.Controls
+{
+    internal static class Helper
+    {
+        public static Differ Instance { get; } = new Differ();
+    }
+}
