+++ DiffPlex.Wpf/Controls/InlineDiffViewer.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="DiffPlex.Wpf.Controls.InlineDiffControl"
+﻿<UserControl x:Class="DiffPlex.Wpf.Controls.InlineDiffViewer"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
@@ -6,5 +6,5 @@
              xmlns:local="clr-namespace:DiffPlex.Wpf.Controls"
              mc:Ignorable="d" 
              d:DesignHeight="450" d:DesignWidth="800">
-    <local:InternalLinesControl x:Name="ContentPanel" />
+    <local:InternalLinesViewer x:Name="ContentPanel" />
 </UserControl>
