+++ DiffPlex.Wpf/Controls/InternalLinesViewer.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="DiffPlex.Wpf.Controls.InternalLinesControl"
+﻿<UserControl x:Class="DiffPlex.Wpf.Controls.InternalLinesViewer"
              x:ClassModifier="internal"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
